// sdk/src/models/message-ref.ts
/*
 * Code generated by Speakeasy (https://speakeasy.com). DO NOT EDIT.
 */

import * as z from "zod/v3";
import { safeParse } from "../lib/schemas.js";
import { Result as SafeParseResult } from "../types/fp.js";
import * as types from "../types/primitives.js";
import { SDKValidationError } from "./errors/sdk-validation-error.js";

/**
 * Lightweight reference returned by write operations.
 */
export type MessageRef = {
  /**
   * Server-assigned message identifier (24-char hex).
   */
  id: string;
  /**
   * ID of the channel that contains the message.
   */
  channelId: string;
};

/** @internal */
export const MessageRef$inboundSchema: z.ZodType<
  MessageRef,
  z.ZodTypeDef,
  unknown
> = z.object({
  id: types.string(),
  channelId: types.string(),
});
/** @internal */
export type MessageRef$Outbound = {
  id: string;
  channelId: string;
};

/** @internal */
export const MessageRef$outboundSchema: z.ZodType<
  MessageRef$Outbound,
  z.ZodTypeDef,
  MessageRef
> = z.object({
  id: z.string(),
  channelId: z.string(),
});

export function messageRefToJSON(messageRef: MessageRef): string {
  return JSON.stringify(MessageRef$outboundSchema.parse(messageRef));
}
export function messageRefFromJSON(
  jsonString: string,
): SafeParseResult<MessageRef, SDKValidationError> {
  return safeParse(
    jsonString,
    (x) => MessageRef$inboundSchema.parse(JSON.parse(x)),
    `Failed to parse 'MessageRef' from JSON`,
  );
}