// sdk/src/extensions/thread-context.ts
export type ThreadContextOptions = RequestOptions;
export type ReplyToThreadOptions = RequestOptions;
// ...
export interface ThreadContextRequest {
  channelId: string;
  messageId: string;
  replyLimit?: number | undefined;
}

export type ReplyToThreadRequest =
  & Omit<SendReplyRequest, "channel" | "channelId" | "messageId">
  & {
    channelId: string;
    messageId: string;
  };